improve code readability and maintainability by adding comments and improving variable names.
