+++ sdk/nexent/core/models/__init__.py
@@ -4,7 +4,7 @@ from .openai_long_context_model import OpenAILongContextModel
 from .stt_model import BaseSTTModel
 from .ali_stt_model import AliSTTModel, AliSTTConfig
 from .volc_stt_model import VolcSTTModel, VolcSTTConfig
-from .base_tts_model import BaseTTSModel
+from .tts_model import BaseTTSModel
 from .ali_tts_model import AliTTSModel, AliTTSConfig
 from .volc_tts_model import VolcTTSModel, VolcTTSConfig
 from . import openai_llm, openai_vlm, openai_long_context_model
