+++ sdk/nexent/core/models/volc_tts_model.py
@@ -12,7 +12,7 @@ from typing import Any, AsyncGenerator, Dict, Optional, Union
 
 import websockets
 
-from .base_tts_model import BaseTTSModel
+from .tts_model import BaseTTSModel
 
 logger = logging.getLogger(__name__)
